HelVM.HelMA.Automata.Piet.LLVM.ImageReader

Documentation

data AdditionalColorStrategy

Constructors

AdditionalColorAsWhite
Treating as a white codel.

AdditionalColorAsBlack
Treating as a black codel.

AdditionalColorNearest
Treating as a codel which has the nearest color.

data CodelSize

Constructors

CodelSize Int

GuessCodelSize

data ImageConfig

Constructors

ImageConfig

Fields

additionalColor :: AdditionalColorStrategy

multicoloredCodel :: MulticoloredCodelStrategy

codelSize :: CodelSize

data ImageReaderError

Constructors

ReadImageFileError String
The image file is unreadable.

UnsupportedImageError String
The input image has an unsupported format.

CodelSizeError
The specified size of codel is not fit for the image.

data MulticoloredCodelStrategy

Constructors

MulticoloredCodelAsWhite
Treating as a white codel.

MulticoloredCodelAsBlack
Treating as a black codel.

MulticoloredCodelCenter
Picking up a center pixel.

MulticoloredCodelModal
Finding the modal color, the most frequent color.

MulticoloredCodelAverage
Calculating an average color.

imageToCodels :: MonadError ImageReaderError m => ImageConfig -> DynamicImage -> m (Vector (Vector Codel))

readCodels :: (MonadIO m, MonadError ImageReaderError m) => ImageConfig -> FilePath -> m (Vector (Vector Codel))

rgbImageToCodels :: MonadError ImageReaderError m => ImageConfig -> Image PixelRGB8 -> m (Vector (Vector Codel))