HelVM.HelMA.Automata.Piet.LLVM.Parser

Documentation

data ParserError

Constructors

EmptyBlockTableError
The block table is empty.

IllegalInitialColorError
The initial codel of the block table is black.

MissingCodelIndexError Int
A codel index in the codel table is missing.

parse :: MonadError ParserError m => Vector (Vector Color) -> m SyntaxGraph

parseFilledImage :: MonadError ParserError m => (CodelTable, BlockTable) -> m SyntaxGraph